change back to 32 bit msg data

+++ tuxbox/neutrino/src/daemonc/remotecontrol.cpp
@@ -82,7 +82,7 @@ CRemoteControl::CRemoteControl()
 	is_video_started = true;
 }
 
-int CRemoteControl::handleMsg(uint msg, uint64_t data)
+int CRemoteControl::handleMsg(uint msg, uint data)
 {
 	if ( zap_completion_timeout != 0 )
 	{
@@ -162,7 +162,7 @@ int CRemoteControl::handleMsg(uint msg, uint64_t data)
 
     if ( msg == NeutrinoMessages::EVT_CURRENTEPG )
 	{
-		CSectionsdClient::CurrentNextInfo* info_CN = (CSectionsdClient::CurrentNextInfo*) (uint)data;
+		CSectionsdClient::CurrentNextInfo* info_CN = (CSectionsdClient::CurrentNextInfo*) data;
 
 		if ( ( info_CN->current_uniqueKey >> 16) == current_channel_id )
 		{
@@ -208,7 +208,7 @@ int CRemoteControl::handleMsg(uint msg, uint64_t data)
 	}
 	else if ( msg == NeutrinoMessages::EVT_NEXTEPG )
 	{
-		CSectionsdClient::CurrentNextInfo* info_CN = (CSectionsdClient::CurrentNextInfo*) (uint)data;
+		CSectionsdClient::CurrentNextInfo* info_CN = (CSectionsdClient::CurrentNextInfo*) data;
 
 		if ( ( info_CN->next_uniqueKey >> 16) == current_channel_id )
 		{
